[#21] Small refactoring for web-service: step 6.2 (network refactoring)
